Moved plugins to subpackages
