+++ core/src/toga/plugins/pil.py
